refactor(rmcp-client): simplify oauth refresh coordination
